Some transitions (Push, Move In, Reveal) leave parts of the slideshow window transparent.  Use a dpf-black-view as a slideshow window's content view to avoid this.  See ticket:41.
Unfortunately, this shows a new problem:  the transition doesn't seem
to respect the rounded corner clipping path.  This looks strange, because
as the new content slides in, it covers the rounded corners as it goes by.

There may be a way to address this by using a CALayer directly, rather
then relying on a layer-backed view hierarchy as I have been doing.